music/pyproject.toml
[tool.poetry]
name = "music"
version = "0.1.0"
description = ""
authors = ["Vadim Bulai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
environ = "^1.0"
django-environ = "^0.11.2"
psycopg2 = "^2.9.7"
django-rest-swagger = "^2.2.0"
django-filter = "^23.2"
djangorestframework-jsonapi = "^6.1.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
